Crate conjure_core

Re-exports

pub extern crate self as conjure_core;

pub use ast::Model;

Modules

ast

bug

context

error

Top-level error types for Conjure-Oxide.

metadata

parse

representation

rule_engine

solver

A high-level API for interacting with constraints solvers.

stats

Macros

bug

Triggers a panic with a detailed bug report message, while ensuring the panic is ignored in coverage reports.

error

into_matrix

Creates a new matrix AbstractLiteral from some Vec, optionally with some index domain.

into_matrix_expr

Creates a new matrix as an Expression from a (Rust) vector, optionally with some index domain.

matrix

Creates a new matrix AbstractLiteral optionally with some index domain.

matrix_expr

Creates a new matrix as an Expression, optionally with some index domain.

register_representation

throw_error