conjure_core

Module ast

Re-exports

pub use model::Model;

pub use types::ReturnType;

Modules

model

pretty

Functions for pretty printing Conjure models.

types

Structs

DecisionVariable

Represents a decision variable within a computational model.

SymbolTable

The global symbol table, mapping names to their definitions.

Enums

Atom

An Atom is an indivisible expression, such as a literal or a reference.

Domain

Expression

Represents different types of expressions used to define rules and constraints in the model.

Literal

A literal value, equivalent to constants in Conjure.

Name

A reference to an object stored in the SymbolTable.

Range