Function conjure_core::rule_engine::get_rule_priorities

pub fn get_rule_priorities<'a>(
    rule_sets: &Vec<&'a RuleSet<'a>>
) -> Result<HashMap<&'a Rule<'a>, u8>, ResolveRulesError>

Convert a list of rule sets into a final map of rules to their priorities.

Arguments

rule_sets The rule sets to get the rules from.

Returns

A map of rules to their priorities.