Macro conjure_core::rule_engine::register_rule_set

register_rule_set!() { /* proc-macro */ }

This procedural macro registers a rule set with the global registry. It may be used in any downstream crate.

For more information on linker magic, see the linkme crate.

This macro uses the following syntax:

register_rule_set!(<RuleSet name>, <RuleSet order>, (<DependencyRuleSet1>, <DependencyRuleSet2>, ...));

Example

use conjure_core::rule_engine::register_rule_set;

register_rule_set!("MyRuleSet", 10, ("DependencyRuleSet", "AnotherRuleSet"));

Register a rule set with the given name, priority, and dependencies.

Example

use conjure_macros::register_rule_set;

register_rule_set!(“MyRuleSet”, 10, (“DependencyRuleSet”, “AnotherRuleSet”));