conjure_core::solver::adaptors::sat_common

Struct CNFModel

pub struct CNFModel {
    pub clauses: Vec<Vec<i32>>,
    /* private fields */
}

A representation of a model in CNF.

Expects Model to be in the Conjunctive Normal Form:

All variables must be boolean

Expressions must be Reference, Not(Reference), or Or(Reference1, Not(Reference2), ...)

The top level And() may contain nested Or()s. Any other nested expressions are not allowed.

Fields
clauses: Vec<Vec<i32>>
Implementations

impl CNFModel

pub fn new() -> CNFModel

pub fn from_conjure( conjure_model: ConjureModel, ) -> Result<CNFModel, SolverError>

pub fn get_variables(&self) -> Vec<&Name>

Gets all the Conjure variables in the CNF.

pub fn get_index(&self, var: &Name) -> Option<i32>

Gets the index of a Conjure variable.

pub fn get_name(&self, ind: i32) -> Option<&Name>

Gets a Conjure variable by index.

pub fn add_variable(&mut self, var: &Name)

Adds a new Conjure variable to the CNF representation.

pub fn has_variable<T: HasVariable>(&self, value: T) -> bool

Check if a Conjure variable or index is present in the CNF

pub fn add_clause(&mut self, vec: &Vec<i32>) -> Result<(), CNFError>

Add a new clause to the CNF. Must be a vector of indices in CNF format

pub fn add_expression(&mut self, expr: &Expression) -> Result<(), CNFError>

Add a new Conjure expression to the CNF. Must be a logical expression in CNF form

pub fn as_expression(&self) -> Result<Expression, CNFError>

Convert the CNF to a Conjure expression