conjure_core::solver
Module model_modifier

Modifying a model during search.

Incremental solving can be triggered for a solverthrough the Solver::solve_mut method.

This gives access to a ModelModifier in the solution retrieval callback.

Structs

NotModifiable

A ModelModifier for a solver that does not support incremental solving. Returns OperationNotSupported for all operations.

Enums

ModificationFailure

The requested modification to the model has failed.

Traits

ModelModifier

A ModelModifier provides an interface to modify a model during solving.