conjure_core::solver
Module states

States of a Solver.

Structs

ExecutionFailure

The state returned by Solver if solving has not been successful.

ExecutionSuccess

The state returned by Solver if solving has been successful.

Init

ModelLoaded

Traits

SolverState