Crate conjure_core

Re-exports

pub extern crate self as conjure_core;

pub use model::Model;

Modules

ast

bug

context

error

Top-level error types for Conjure-Oxide.

metadata

model

parse

rule_engine

rules

This module contains the rewrite rules for Conjure Oxides and it’s solvers.

solver

A high-level API for interacting with constraints solvers.

stats

Macros

bug

Triggers a panic with a detailed bug report message, while ensuring the panic is ignored in coverage reports.