conjure_core
Module ast

Re-exports

pub use types::ReturnType;

Modules

types

Structs

DecisionVariable

Represents a decision variable within a computational model.

Enums

Atom

An Atom is an indivisible expression, such as a literal or a reference.

Domain

Expression

Represents different types of expressions used to define rules and constraints in the model.

Literal

A literal value, equivalent to constants in Conjure.

Name

Range

Type Aliases

SymbolTable